JPM: added check for JCC 3.0

+++ src_pylucene/build.sh
@@ -36,22 +36,36 @@ if [ -z "$libDir" ]; then
     libDir=$(dirname $mydir)/lib
 fi
 
-rm -rf $buildDir $libDir
-mkdir --parents $buildDir $libDir
-
 pythonVersion=$(python --version 2>&1 | awk '{print $2}' | cut -d. -f-2)
+pythonPackagesDir=/usr/lib64/python${pythonVersion}/site-packages
+if [ -f /etc/debian_version ]; then
+    pythonPackagesDir=/usr/lib/python${pythonVersion}/dist-packages
+fi
+
+JCC_VERSION=3.0
+if ! grep -q "VERSION=\"${JCC_VERSION}\"" ${pythonPackagesDir}/jcc/config.py; then
+    echo "JCC ${JCC_VERSION} is required."
+    exit 1
+fi
 
 JAVA_VERSION=8
 javac=/usr/lib/jvm/java-1.${JAVA_VERSION}.0/bin/javac
-luceneJarDir=/usr/lib64/python${pythonVersion}/site-packages/lucene
 if [ -f /etc/debian_version ]; then
     javac=/usr/lib/jvm/java-${JAVA_VERSION}-openjdk-amd64/bin/javac
-    luceneJarDir=/usr/lib/python${pythonVersion}/dist-packages/lucene
 fi
+if [ ! -f "$javac" ]; then
+    echo "No Java ${JAVA_VERSION} javac found."
+    exit 1
+fi
+
+luceneJarDir=${pythonPackagesDir}/lucene
 
 PYLUCENEVERSION=6.5.0
 
 classpath=${luceneJarDir}/lucene-core-$PYLUCENEVERSION.jar:${luceneJarDir}/lucene-analyzers-common-$PYLUCENEVERSION.jar
+
+rm -rf $buildDir $libDir
+mkdir --parents $buildDir $libDir
 
 ${javac} -cp ${classpath} -d ${buildDir} `find . -name "*.java"`
 (cd $buildDir; jar -c org > $buildDir/meresco-lucene.jar)
@@ -75,4 +89,3 @@ fi
 mv ${rootLibDir} $libDir/
 
 rm -rf $buildDir $mydir/root $mydir/meresco_lucene.egg-info
-
